First pass of modified imports, move deprecated classes to ideas/deprecated
