+++ ideas/deprecated/NOTES-ABOUT-OBSERVATION-CLASSES.txt
