+++ ideas/deprecated/extendedsource.py
