+++ ideas/deprecated/pointsource.py
