+++ ideas/deprecated/rasterscan.py
