+++ ideas/deprecated/rasterslit.py
