+++ ideas/deprecated/slit.py
