+++ oops/calibration/all.py
@@ -5,10 +5,8 @@
 # Import the Calibration class and all its subclasses into a common name space
 
 from oops.calibration                import Calibration
-from oops.calibration.extendedsource import ExtendedSource  # DEPRECATED
 from oops.calibration.flatcalib      import FlatCalib
 from oops.calibration.nullcalib      import NullCalib
-from oops.calibration.pointsource    import PointSource     # DEPRECATED
 from oops.calibration.radiance       import Radiance
 from oops.calibration.rawcounts      import RawCounts
