+++ oops/calibration/unittester.py
@@ -5,10 +5,8 @@
 import unittest
 
 from oops.calibration                import Test_Calibration
-from oops.calibration.extendedsource import Test_ExtendedSource # DEPRECATED
 from oops.calibration.flatcalib      import Test_FlatCalib
 from oops.calibration.nullcalib      import Test_NullCalib
-from oops.calibration.pointsource    import Test_PointSource    # DEPRECATED
 from oops.calibration.radiance       import Test_Radiance
 from oops.calibration.rawcounts      import Test_RawCounts
