+++ oops/fov/all.py
@@ -9,9 +9,7 @@ from oops.fov.barrelfov     import BarrelFOV
 from oops.fov.flatfov       import FlatFOV
 from oops.fov.nullfov       import NullFOV
 from oops.fov.offsetfov     import OffsetFOV
-from oops.fov.polyfov       import PolyFOV
 from oops.fov.polynomialfov import PolynomialFOV
-from oops.fov.radialfov     import RadialFOV
 from oops.fov.slicefov      import SliceFOV
 from oops.fov.subarray      import Subarray
 from oops.fov.subsampledfov import SubsampledFOV
