+++ oops/fov/unittester.py
@@ -8,9 +8,7 @@ from oops.fov               import Test_FOV
 from oops.fov.flatfov       import Test_FlatFOV
 from oops.fov.barrelfov     import Test_BarrelFOV
 from oops.fov.offsetfov     import Test_OffsetFOV
-from oops.fov.polyfov       import Test_PolyFOV
 from oops.fov.polynomialfov import Test_PolynomialFOV
-from oops.fov.radialfov     import Test_RadialFOV
 from oops.fov.slicefov      import Test_SliceFOV
 from oops.fov.subarray      import Test_Subarray
 from oops.fov.subsampledfov import Test_SubsampledFOV
